user_org_service/requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
boto3==1.42.34
botocore==1.42.34
celery==5.6.2
certifi==2026.1.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.13.2
cron-descriptor==2.0.6
django==5.2.10
django-celery-beat==2.8.1
django-storages==1.14.6
django-timezone-field==7.2.1
djangorestframework==3.16.1
iniconfig==2.3.0
isort==7.0.0
jmespath==1.1.0
kombu==5.6.2
librt==0.7.8
mypy==1.19.1
mypy-extensions==1.1.0
packaging==26.0
pathspec==1.0.3
pika==1.3.2
pluggy==1.6.0
prompt-toolkit==3.0.52
psycopg2-binary==2.9.11
pygments==2.19.2
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
redis==7.1.0
ruff==0.14.14
s3transfer==0.16.0
sentry-sdk==2.50.0
six==1.17.0
sqlparse==0.5.5
typing-extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.3.5
